+++ src/main/scala/ai/metarank/fstore/redis/RedisPersistence.scala
@@ -1,30 +1,13 @@
 package ai.metarank.fstore.redis
 
-import ai.metarank.config.StateStoreConfig.RedisCredentials
+import ai.metarank.config.StateStoreConfig.{RedisCredentials, RedisTLS, RedisTimeouts}
 import ai.metarank.config.StateStoreConfig.RedisStateConfig.{CacheConfig, DBConfig, PipelineConfig}
-import ai.metarank.fstore.cache.CachedFeature.{
-  CachedBoundedListFeature,
-  CachedCounterFeature,
-  CachedFreqEstimatorFeature,
-  CachedMapFeature,
-  CachedPeriodicCounterFeature,
-  CachedScalarFeature,
-  CachedStatsEstimatorFeature
-}
+import ai.metarank.fstore.cache.CachedFeature.{CachedBoundedListFeature, CachedCounterFeature, CachedFreqEstimatorFeature, CachedMapFeature, CachedPeriodicCounterFeature, CachedScalarFeature, CachedStatsEstimatorFeature}
 import ai.metarank.fstore.Persistence
 import ai.metarank.fstore.Persistence.{KVCodec, ModelName}
 import ai.metarank.fstore.cache.{CachedClickthroughStore, CachedKVStore}
 import ai.metarank.fstore.codec.StoreFormat
-import ai.metarank.fstore.memory.{
-  MemBoundedList,
-  MemCounter,
-  MemFreqEstimator,
-  MemKVStore,
-  MemMapFeature,
-  MemPeriodicCounter,
-  MemScalarFeature,
-  MemStatsEstimator
-}
+import ai.metarank.fstore.memory.{MemBoundedList, MemCounter, MemFreqEstimator, MemKVStore, MemMapFeature, MemPeriodicCounter, MemScalarFeature, MemStatsEstimator}
 import ai.metarank.fstore.redis.client.RedisClient
 import ai.metarank.model.{FeatureValue, Key, Schema}
 import ai.metarank.rank.Model.Scorer
@@ -163,7 +146,7 @@ case class RedisPersistence(
     _     <- stateClient.doFlush(stateClient.writer.ping().toCompletableFuture)
     _     <- valuesClient.doFlush(valuesClient.writer.ping().toCompletableFuture)
 //    _     <- rankingsClient.doFlush(rankingsClient.writer.ping().toCompletableFuture)
-    _     <- modelClient.doFlush(modelClient.writer.ping().toCompletableFuture)
+    _ <- modelClient.doFlush(modelClient.writer.ping().toCompletableFuture)
   } yield {
     logger.info(s"redis pipeline flushed, took ${System.currentTimeMillis() - start}ms")
   }
@@ -184,12 +167,13 @@ object RedisPersistence {
       cache: CacheConfig,
       pipeline: PipelineConfig,
       format: StoreFormat,
-      auth: Option[RedisCredentials]
+      auth: Option[RedisCredentials],
+      tls: Option[RedisTLS],
+      timeout: RedisTimeouts
   ): Resource[IO, RedisPersistence] = for {
-    state    <- RedisClient.create(host, port, db.state, pipeline, auth)
-    models   <- RedisClient.create(host, port, db.models, pipeline, auth)
-    values   <- RedisClient.create(host, port, db.values, pipeline, auth)
-//    rankings <- RedisClient.create(host, port, db.rankings, pipeline, auth)
+    state  <- RedisClient.create(host, port, db.state, pipeline, auth, tls,timeout)
+    models <- RedisClient.create(host, port, db.models, pipeline, auth, tls,timeout)
+    values <- RedisClient.create(host, port, db.values, pipeline, auth, tls,timeout)
     _ <- Resource.liftK(
       IO.fromCompletableFuture(
         IO(