+++ doc/changelog.md
@@ -3,6 +3,11 @@
 In a human-readable format. For a technical changelog for robots, see [github releases page](https://github.com/metarank/metarank/releases).
 Check our [blog](https://blog.metarank.ai) for more detailed updates.
 
+## 0.5.10
+
+* Redis [TLS support](configuration/persistence.md#tls-support)
+* Redis [timeout configuration](configuration/overview.md#persistence)
+
 ## 0.5.9
 
 * [`interacted_with`](configuration/features/user-session.md#interacted-with) feature now has much less overhead in Redis, and supports multiple fields in a single visitor profile.